Adding test coverage for Model.FindElementByPath
There are two failing tests at this point, because collections define
their path to be equal to their parent's path, so all calls to this method
for collections will return the Model instead of the specific item.

This behavior cannot be fixed right now, but it should be something to
think about in the future.